requirements.txt
flask
numpy
opencv-python
pillow
onnxruntime
gunicorn
